# gateway/src/cities/cities.controller.ts
import {Body, Controller, Delete, Get, Param, Patch, Post} from '@nestjs/common';
import {CitiesService} from './cities.service';
import {CreateCityDto} from './dto/create-city.dto';
import {UpdateCityDto} from './dto/update-city.dto';

@Controller('cities')
export class CitiesController {
    constructor(private readonly citiesService: CitiesService) {
    }

    @Post()
    create(@Body() createCityDto: CreateCityDto) {
        return this.citiesService.create(createCityDto);
    }

    @Get()
    findAll() {
        return this.citiesService.findAll();
    }

    @Get(':id')
    findOne(@Param('id') id: string) {
        return this.citiesService.findOne(+id);
    }

    @Patch(':id')
    update(@Param('id') id: string, @Body() updateCityDto: UpdateCityDto) {
        return this.citiesService.update(+id, updateCityDto);
    }

    @Delete(':id')
    remove(@Param('id') id: string) {
        return this.citiesService.remove(+id);
    }
}