# gateway/src/countries/countries.controller.ts
import {Body, Controller, Delete, Get, Param, Patch, Post} from '@nestjs/common';
import {CountriesService} from './countries.service';
import {CreateCountryDto} from './dto/create-country.dto';
import {UpdateCountryDto} from './dto/update-country.dto';

@Controller('countries')
export class CountriesController {
    constructor(private readonly countriesService: CountriesService) {
    }

    @Post()
    create(@Body() createCountryDto: CreateCountryDto) {
        return this.countriesService.create(createCountryDto);
    }

    @Get()
    findAll() {
        return this.countriesService.findAll();
    }

    @Get(':id')
    findOne(@Param('id') id: string) {
        return this.countriesService.findOne(+id);
    }

    @Patch(':id')
    update(@Param('id') id: string, @Body() updateCountryDto: UpdateCountryDto) {
        return this.countriesService.update(+id, updateCountryDto);
    }

    @Delete(':id')
    remove(@Param('id') id: string) {
        return this.countriesService.remove(+id);
    }
}