# gateway/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/users":{"get":{"operationId":"UsersController_all","parameters":[{"name":"offset","required":true,"in":"query","schema":{"default":0,"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"default":10,"type":"number"}},{"name":"orderBy","required":false,"in":"query","schema":{"type":"string"}},{"name":"orderType","required":false,"in":"query","schema":{"default":"ASC","type":"string"}},{"name":"firstName","required":false,"in":"query","schema":{"type":"string"}},{"name":"lastName","required":false,"in":"query","schema":{"type":"string"}},{"name":"email","required":false,"in":"query","schema":{"type":"string"}},{"name":"password","required":false,"in":"query","schema":{"type":"string"}},{"name":"role","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get all users","tags":["Users"]},"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"201":{"description":""}},"summary":"Create User","tags":["Users"]}},"/api/users/{id}":{"get":{"operationId":"UsersController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Get User by ID","tags":["Users"]},"patch":{"operationId":"UsersController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"200":{"description":""}},"summary":"Update User","tags":["Users"]},"delete":{"operationId":"UsersController_delete","parameters":[],"responses":{"200":{"description":""}},"summary":"Delete User","tags":["Users"]}},"/api/users/whoami":{"get":{"operationId":"UsersController_whoami","parameters":[],"responses":{"200":{"description":""}},"summary":"Get User by ID","tags":["Users"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/roles/{id}":{"get":{"operationId":"RolesController_getById","parameters":[],"responses":{"200":{"description":""}},"summary":"Get Role by ID","tags":["Roles"]},"patch":{"operationId":"RolesController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleDto"}}}},"responses":{"200":{"description":""}},"summary":"Update Role","tags":["Roles"]},"delete":{"operationId":"RolesController_delete","parameters":[],"responses":{"200":{"description":""}},"summary":"Delete Role","tags":["Roles"]}},"/api/roles":{"get":{"operationId":"RolesController_all","parameters":[{"name":"offset","required":true,"in":"query","schema":{"default":0,"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"default":10,"type":"number"}},{"name":"orderBy","required":false,"in":"query","schema":{"type":"string"}},{"name":"orderType","required":false,"in":"query","schema":{"default":"ASC","type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get all roles","tags":["Roles"]},"post":{"operationId":"RolesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleDto"}}}},"responses":{"201":{"description":""}},"summary":"Create Role","tags":["Roles"]}},"/api/clients":{"post":{"operationId":"ClientsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientDto"}}}},"responses":{"201":{"description":"The client has been successfully created."},"400":{"description":"Bad Request."}},"summary":"Create a new client","tags":["clients"]},"get":{"operationId":"ClientsController_findAll","parameters":[],"responses":{"200":{"description":"Return all clients."}},"summary":"Get all clients","tags":["clients"]}},"/api/clients/{id}":{"get":{"operationId":"ClientsController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"Client ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Return the client."},"404":{"description":"Client not found."}},"summary":"Get a client by ID","tags":["clients"]},"patch":{"operationId":"ClientsController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Client ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateClientDto"}}}},"responses":{"200":{"description":"The client has been successfully updated."},"404":{"description":"Client not found."}},"summary":"Update a client by ID","tags":["clients"]},"delete":{"operationId":"ClientsController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"Client ID","schema":{"type":"string"}}],"responses":{"200":{"description":"The client has been successfully deleted."},"404":{"description":"Client not found."}},"summary":"Delete a client by ID","tags":["clients"]}},"/api/cities":{"post":{"operationId":"CitiesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCityDto"}}}},"responses":{"201":{"description":""}},"tags":["Cities"]},"get":{"operationId":"CitiesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Cities"]}},"/api/cities/{id}":{"get":{"operationId":"CitiesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Cities"]},"patch":{"operationId":"CitiesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCityDto"}}}},"responses":{"200":{"description":""}},"tags":["Cities"]},"delete":{"operationId":"CitiesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Cities"]}},"/api/states":{"post":{"operationId":"StatesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStateDto"}}}},"responses":{"201":{"description":""}},"tags":["States"]},"get":{"operationId":"StatesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["States"]}},"/api/states/{id}":{"get":{"operationId":"StatesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["States"]},"patch":{"operationId":"StatesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStateDto"}}}},"responses":{"200":{"description":""}},"tags":["States"]},"delete":{"operationId":"StatesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["States"]}},"/api/countries":{"post":{"operationId":"CountriesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCountryDto"}}}},"responses":{"201":{"description":""}},"tags":["Countries"]},"get":{"operationId":"CountriesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Countries"]}},"/api/countries/{id}":{"get":{"operationId":"CountriesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Countries"]},"patch":{"operationId":"CountriesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCountryDto"}}}},"responses":{"200":{"description":""}},"tags":["Countries"]},"delete":{"operationId":"CountriesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Countries"]}}},"info":{"title":"FMA API","description":"","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"api_key":{"type":"apiKey","in":"header","name":"x-api-key","description":"API Key For External calls"},"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"apiKey","name":"Authorization","in":"header","description":"JWT Token use Bearer"}},"schemas":{"UserDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}}},"RoleDto":{"type":"object","properties":{}},"CreateClientDto":{"type":"object","properties":{}},"UpdateClientDto":{"type":"object","properties":{}},"CreateCityDto":{"type":"object","properties":{}},"UpdateCityDto":{"type":"object","properties":{}},"CreateStateDto":{"type":"object","properties":{}},"UpdateStateDto":{"type":"object","properties":{}},"CreateCountryDto":{"type":"object","properties":{}},"UpdateCountryDto":{"type":"object","properties":{}}}}}